Migration browsefiles to django
